move PickerDialog to SelectBoardLayoutDialog
