--- src/modules/ui/picker-dialog/index.ts
@@ -1,3 +1,0 @@
-import PickerDialog from "./PickerDialog.vue";
-
-export { PickerDialog };
