--- src/modules/ui/picker-dialog/types.ts
@@ -1,7 +1,0 @@
-export type PickerOption = {
-	icon: string;
-	label: string;
-	action: () => void;
-	dataTestId: string;
-	ariaLabel: string;
-};
