+++ src/modules/ui/room-details/SelectBoardLayoutDialog.vue
